Fix to #8315 - Argument types don't match for queries projecting conditional expression with anonymous type result
Problem was that when translating anonymous type we change the expression type from the anonymous type to Expression[].
In case of conditional expression, if the second result is null constant, it's type stays the same, and when we try to update ConditionalExpression, type mismatch is thrown.

Fix is to recognize the pattern and force client eval for this case, since we cant map objects of type Expression[] anyway.