Fix to #7975 - Query: we should propagate optional navigations thru nav rewrite, if the initial optional navigation was created using SelectMany-GroupJoin-DefaultIfEmpty
also fix to #7761 - Query : Missing null propagation logic for some complex queries with manually created GroupJoin-SelectMany-DefaultIfEmpty

Problem for queries with required navigations that were originating from qsre that was a GroupJoin-SelectMany-DefaultIfEmpty. In those cases, we should be producing LEFT JOINs for the navigations, even though they are required (because initial qsre is already "optional"). We would produce INNER JOINs instead, which would result in data corruption - returning fewer results than expected.

Fix is to use more sophisticated logic to determine whether the source qsre is "optional" and therefore we need to expand all its navigations to LOJs.

We also apply this logic to MemberAccess and EF.Property, effectively fully protecting against accidental NREs for InMemory and client eval scenarios (we were doing that before only for results of nav rewrite, now we also do it for manually created LOJs).